requirements.txt
aniso8601==7.0.0
attrs==19.1.0
backpack==0.1
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
cleo==0.6.8
Click==7.0
clikit==0.3.1
Faker==0.8.18
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
flask-orator==0.2.0
flask-restplus==0.13.0
Flask-Script==2.0.6
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2==2.8.3
pyaml==16.12.2
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylev==1.3.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
pytzdata==2019.2
PyYAML==5.1.2
simplejson==3.16.0
six==1.12.0
text-unidecode==1.2
tzlocal==1.5.1
Werkzeug==0.15.6
wrapt==1.11.2
